requirements.txt
pandas
requests
python-telegram-bot
flask
